refactor: split GTK top bar into focused modules
